pyproject.toml
[tool.poetry]
name = "django-training"
version = "0.1.0"
description = ""
authors = ["mohamed khaled <moham35356@gmail.com>"]
readme = "README.md"
packages = [{include = "django_training"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
django-model-utils = "^4.2.0"
django-imagekit = "^4.1.0"
pillow = "^9.2.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
django-extensions = "^3.2.1"
django-rest-knox = "^4.2.0"
pytest-django = "^4.5.2"
celery = "^5.2.7"
django-celery-beat = "^2.4.0"
redis = "^4.3.4"
python-crontab = "^2.6.0"
django-environ = "^0.9.0"
django-celery-results = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
